pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "leannvault"
version = "0.1.0"
description = "Hash-based vector search with LEANN - keep your index valid when files move"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "LeannVault Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "leann>=0.3.0",
    "click>=8.0.0",
    "python-docx>=0.8.11",
    "python-pptx>=0.6.21",
    "pdfplumber>=0.10.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "gradio>=4.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
leannvault = "leannvault.cli.main:cli"

[project.urls]
Homepage = "https://github.com/leannvault/leannvault"
Documentation = "https://github.com/leannvault/leannvault#readme"
Repository = "https://github.com/leannvault/leannvault"

[tool.hatch.build.targets.wheel]
packages = ["src/leannvault"]

[tool.hatch.build.targets.sdist]
include = ["src/leannvault", "README.md", "pyproject.toml"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
